+++ packages/shared-ui/src/lib-examples/simple-color-switch/index.tsx
@@ -1,3 +1,0 @@
-export default function SimpleColorSwitch() {
-  return <div></div>;
-}