+++ packages/shared-ui/src/lib-examples/theme-switching/index.tsx
@@ -1,3 +1,0 @@
-export default function ThemeSwitching() {
-  return <div></div>;
-}